telebot/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
behave==1.2.6
cachier==1.5.3
certifi==2023.5.7
charset-normalizer==2.0.12
click==7.1
colorama==0.4.6
dispatcher==1.0
Django==4.2.2
django-admin-list-filter-dropdown==1.0.3
django-filter==23.2
django-json-widget==1.1.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
loguru==0.5.3
magic-filter==1.0.9
Markdown==3.4.4
MarkupSafe==2.1.3
multidict==6.0.4
openpyxl==3.1.2
packaging==21.3
parse==1.19.1
parse-type==0.6.0
pathtools==0.1.2
Pillow==9.5.0
plumbum==1.7.1
portalocker==2.7.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pymongo==3.12.1
pyparsing==3.1.0
pypiwin32==223
pyTelegramBotAPI==4.12.0
python-rex==0.4
python-telegram-bot==20.3
pytz==2023.3
pywin32==306
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
telebot==0.0.5
telegram-django-bot==1.1.0
tzdata==2023.3
updater==0.3.2
urllib3==1.26.16
watchdog==3.0.0
win32-setctime==1.1.0
yarl==1.9.2